improve comments & general cleanup
